moved beacon's gateway to shared and moved the main func to its own /server folder
