--- beacon-chain/gateway/BUILD.bazel
@@ -1,29 +1,0 @@
-# gazelle:ignore
-load("@io_bazel_rules_go//go:def.bzl", "go_library")
-
-go_library(
-    name = "go_default_library",
-    srcs = [
-        "cors.go",
-        "gateway.go",
-        "handlers.go",
-        "log.go",
-    ],
-    importpath = "github.com/prysmaticlabs/prysm/beacon-chain/gateway",
-    visibility = [
-        "//beacon-chain/gateway/server:__pkg__",
-        "//beacon-chain/node:__pkg__",
-    ],
-    deps = [
-        "//proto/beacon/rpc/v1:go_grpc_gateway_library",
-        "//shared:go_default_library",
-        "@com_github_grpc_ecosystem_grpc_gateway//runtime:go_default_library",
-        "@com_github_pkg_errors//:go_default_library",
-        "@com_github_prysmaticlabs_ethereumapis//eth/v1alpha1:go_grpc_gateway_library",
-        "@com_github_rs_cors//:go_default_library",
-        "@com_github_sirupsen_logrus//:go_default_library",
-        "@org_golang_google_grpc//:go_default_library",
-        "@org_golang_google_grpc//connectivity:go_default_library",
-        "@org_golang_google_grpc//credentials:go_default_library",
-    ],
-)
