--- beacon-chain/gateway/cors.go
@@ -1,20 +1,0 @@
-package gateway
-
-import (
-	"net/http"
-
-	"github.com/rs/cors"
-)
-
-func newCorsHandler(srv http.Handler, allowedOrigins []string) http.Handler {
-	if len(allowedOrigins) == 0 {
-		return srv
-	}
-	c := cors.New(cors.Options{
-		AllowedOrigins: allowedOrigins,
-		AllowedMethods: []string{http.MethodPost, http.MethodGet},
-		MaxAge:         600,
-		AllowedHeaders: []string{"*"},
-	})
-	return c.Handler(srv)
-}
