--- beacon-chain/gateway/log.go
@@ -1,5 +1,0 @@
-package gateway
-
-import "github.com/sirupsen/logrus"
-
-var log = logrus.WithField("prefix", "gateway")
