--- beacon-chain/gateway/server/BUILD.bazel
@@ -1,50 +1,0 @@
-load("@prysm//tools/go:def.bzl", "go_library")
-load("@io_bazel_rules_go//go:def.bzl", "go_binary")
-load("@io_bazel_rules_docker//go:image.bzl", "go_image")
-load("@io_bazel_rules_docker//container:container.bzl", "container_bundle")
-load("@io_bazel_rules_docker//contrib:push-all.bzl", "docker_push")
-
-go_library(
-    name = "go_default_library",
-    srcs = [
-        "log.go",
-        "main.go",
-    ],
-    importpath = "github.com/prysmaticlabs/prysm/beacon-chain/gateway/server",
-    visibility = ["//visibility:private"],
-    deps = [
-        "//beacon-chain/gateway:go_default_library",
-        "//shared/maxprocs:go_default_library",
-        "@com_github_joonix_log//:go_default_library",
-        "@com_github_sirupsen_logrus//:go_default_library",
-    ],
-)
-
-go_binary(
-    name = "server",
-    embed = [":go_default_library"],
-    visibility = ["//visibility:private"],
-)
-
-go_image(
-    name = "image",
-    base = "//tools:go_image",
-    binary = ":server",
-    tags = ["manual"],
-    visibility = ["//visibility:private"],
-)
-
-container_bundle(
-    name = "image_bundle",
-    images = {
-        "gcr.io/prysmaticlabs/prysm/beacon-chain/gateway:latest": ":image",
-        "gcr.io/prysmaticlabs/prysm/beacon-chain/gateway:{DOCKER_TAG}": ":image",
-    },
-    tags = ["manual"],
-)
-
-docker_push(
-    name = "push_images",
-    bundle = ":image_bundle",
-    tags = ["manual"],
-)
