+++ beacon-chain/node/BUILD.bazel
@@ -21,7 +21,6 @@ go_library(
         "//beacon-chain/db/kv:go_default_library",
         "//beacon-chain/forkchoice:go_default_library",
         "//beacon-chain/forkchoice/protoarray:go_default_library",
-        "//beacon-chain/gateway:go_default_library",
         "//beacon-chain/interop-cold-start:go_default_library",
         "//beacon-chain/node/registration:go_default_library",
         "//beacon-chain/operations/attestations:go_default_library",
@@ -40,6 +39,7 @@ go_library(
         "//shared/debug:go_default_library",
         "//shared/event:go_default_library",
         "//shared/featureconfig:go_default_library",
+        "//shared/gateway:go_default_library",
         "//shared/params:go_default_library",
         "//shared/prereq:go_default_library",
         "//shared/prometheus:go_default_library",
