+++ beacon-chain/server/BUILD.bazel
@@ -1,0 +1,24 @@
+load("@io_bazel_rules_go//go:def.bzl", "go_binary")
+load("@prysm//tools/go:def.bzl", "go_library")
+
+go_library(
+    name = "go_default_library",
+    srcs = [
+        "log.go",
+        "main.go",
+    ],
+    importpath = "github.com/prysmaticlabs/prysm/beacon-chain/server",
+    visibility = ["//visibility:private"],
+    deps = [
+        "//shared/gateway:go_default_library",
+        "//shared/maxprocs:go_default_library",
+        "@com_github_joonix_log//:go_default_library",
+        "@com_github_sirupsen_logrus//:go_default_library",
+    ],
+)
+
+go_binary(
+    name = "server",
+    embed = [":go_default_library"],
+    visibility = ["//visibility:public"],
+)
