+++ shared/gateway/BUILD.bazel
@@ -9,14 +9,17 @@ go_library(
     importpath = "github.com/prysmaticlabs/prysm/shared/gateway",
     visibility = ["//visibility:public"],
     deps = [
+        "//proto/beacon/rpc/v1:go_grpc_gateway_library",
         "//proto/validator/accounts/v2:ethereum_validator_account_gateway_proto",
         "//shared:go_default_library",
         "//validator/web:go_default_library",
         "@com_github_grpc_ecosystem_grpc_gateway//runtime:go_default_library",
         "@com_github_pkg_errors//:go_default_library",
+        "@com_github_prysmaticlabs_ethereumapis//eth/v1alpha1:go_grpc_gateway_library",
         "@com_github_rs_cors//:go_default_library",
         "@com_github_sirupsen_logrus//:go_default_library",
         "@org_golang_google_grpc//:go_default_library",
         "@org_golang_google_grpc//connectivity:go_default_library",
+        "@org_golang_google_grpc//credentials:go_default_library",
     ],
 )
